fix: patches(remove_useless_lib.sh): do not remove `libvvb2060.so` in newer app version `v6.5.7+`
removing it in v6.5.7+ trigger checksum failing error (GW934 - Re-install app to continue using app)

+++ MBCPApp/patches/remove_useless_lib.sh
@@ -5,8 +5,11 @@
 # shellcheck disable=SC1091
 . "$DIRPATH"/common.sh
 
+# remove rootbeer simple check (used in singalarity ekyc)
 rm -rf 'mbapk/mbapk_unpacked/root/lib/arm64-v8a/libtoolChecker.so'
 rm -rf 'mbapk/mbapk_unpacked/root/lib/armeabi-v7a/libtoolChecker.so'
 
-rm -rf 'mbapk/mbapk_unpacked/root/lib/arm64-v8a/libvvb2060.so'
-rm -rf 'mbapk/mbapk_unpacked/root/lib/armeabi-v7a/libvvb2060.so'
+# remove libvvb2060.so (only for older app verion : v6.5.6 or older)
+# in newer app, removing it will trigger checksum failing error (GW934 - Re-install app to continue using app)
+[ ! -f 'mbapk/mbapk_unpacked/root/lib/arm64-v8a/libcode.so' ] && rm -rf 'mbapk/mbapk_unpacked/root/lib/arm64-v8a/libvvb2060.so'
+[ ! -f 'mbapk/mbapk_unpacked/root/lib/armeabi-v7a/libcode.so' ] && rm -rf 'mbapk/mbapk_unpacked/root/lib/armeabi-v7a/libvvb2060.so'
